+++ src/build.ts
@@ -3,12 +3,11 @@ import { promises as fsp } from "node:fs";
 import { resolve, relative, basename } from "pathe";
 import type { PackageJson } from "pkg-types";
 import chalk from "chalk";
-import consola from "consola";
+import { consola } from "consola";
 import { defu } from "defu";
 import { createHooks } from "hookable";
 import prettyBytes from "pretty-bytes";
 import { globby } from "globby";
-import mkdirp from "mkdirp";
 import { dumpObject, rmdir, tryRequire, resolvePreset } from "./utils";
 import type { BuildContext, BuildConfig, BuildOptions } from "./types";
 import { validatePackage, validateDependencies } from "./validate";
@@ -16,62 +15,97 @@ import { rollupBuild } from "./builder/rollup";
 import { typesBuild } from "./builder/untyped";
 import { mkdistBuild } from "./builder/mkdist";
 
-export async function build (rootDir: string, stub: boolean, inputConfig: BuildConfig = {}) {
+export async function build(
+  rootDir: string,
+  stub: boolean,
+  inputConfig: BuildConfig = {}
+) {
   // Determine rootDir
   rootDir = resolve(process.cwd(), rootDir || ".");
 
-  // Read build.config and package.json
-  const buildConfig: BuildConfig = tryRequire("./build.config", rootDir) || {};
-  const pkg: PackageJson & Record<"unbuild" | "build", BuildConfig> = tryRequire("./package.json", rootDir);
+  const _buildConfig: BuildConfig | BuildConfig[] =
+    tryRequire("./build.config", rootDir) || {};
+  const buildConfigs = (
+    Array.isArray(_buildConfig) ? _buildConfig : [_buildConfig]
+  ).filter(Boolean);
 
+  const pkg: PackageJson & Record<"unbuild" | "build", BuildConfig> =
+    tryRequire("./package.json", rootDir) || {};
+
+  // Invoke build for every build config defined in build.config.ts
+  for (const buildConfig of buildConfigs) {
+    await _build(rootDir, stub, inputConfig, buildConfig, pkg);
+  }
+}
+
+async function _build(
+  rootDir: string,
+  stub: boolean,
+  inputConfig: BuildConfig = {},
+  buildConfig: BuildConfig,
+  pkg: PackageJson & Record<"unbuild" | "build", BuildConfig>
+) {
   // Resolve preset
-  const preset = resolvePreset(buildConfig.preset || pkg.unbuild?.preset || pkg.build?.preset || inputConfig.preset || "auto", rootDir);
+  const preset = resolvePreset(
+    buildConfig.preset ||
+      pkg.unbuild?.preset ||
+      pkg.build?.preset ||
+      inputConfig.preset ||
+      "auto",
+    rootDir
+  );
 
   // Merge options
-  const options = defu(buildConfig, pkg.unbuild || pkg.build, inputConfig, preset, <BuildOptions>{
-    name: (pkg?.name || "").split("/").pop() || "default",
-    rootDir,
-    entries: [],
-    clean: true,
-    declaration: false,
-    outDir: "dist",
-    stub,
-    externals: [
-      ...Module.builtinModules,
-      ...Module.builtinModules.map(m => "node:" + m)
-    ],
-    dependencies: [],
-    devDependencies: [],
-    peerDependencies: [],
-    alias: {},
-    replace: {},
-    failOnWarn: true,
-    rollup: {
-      emitCJS: false,
-      cjsBridge: false,
-      inlineDependencies: false,
-      // Plugins
-      replace: {
-        preventAssignment: true
-      },
+  const options = defu(
+    buildConfig,
+    pkg.unbuild || pkg.build,
+    inputConfig,
+    preset,
+    <BuildOptions>{
+      name: (pkg?.name || "").split("/").pop() || "default",
+      rootDir,
+      entries: [],
+      clean: true,
+      declaration: false,
+      outDir: "dist",
+      stub,
+      externals: [
+        ...Module.builtinModules,
+        ...Module.builtinModules.map((m) => "node:" + m),
+      ],
+      dependencies: [],
+      devDependencies: [],
+      peerDependencies: [],
       alias: {},
-      resolve: {
-        preferBuiltins: true
+      replace: {},
+      failOnWarn: true,
+      rollup: {
+        emitCJS: false,
+        cjsBridge: false,
+        inlineDependencies: false,
+        // Plugins
+        replace: {
+          preventAssignment: true,
+        },
+        alias: {},
+        resolve: {
+          preferBuiltins: true,
+        },
+        json: {
+          preferConst: true,
+        },
+        commonjs: {
+          ignoreTryCatch: true,
+        },
+        esbuild: { target: "es2020" },
+        dts: {
+          // https://github.com/Swatinem/rollup-plugin-dts/issues/143
+          compilerOptions: { preserveSymlinks: false },
+          respectExternal: true,
+        },
       },
-      json: {
-        preferConst: true
-      },
-      commonjs: {
-        ignoreTryCatch: true
-      },
-      esbuild: { target: "es2020" },
-      dts: {
-        // https://github.com/Swatinem/rollup-plugin-dts/issues/143
-        compilerOptions: { preserveSymlinks: false },
-        respectExternal: true
-      }
     }
-  }) as BuildOptions;
+  ) as BuildOptions;
 
   // Resolve dirs relative to rootDir
   options.outDir = resolve(options.rootDir, options.outDir);
@@ -83,7 +117,7 @@ export async function build (rootDir: string, stub: boolean, inputConfig: BuildC
     pkg,
     buildEntries: [],
     usedImports: new Set(),
-    hooks: createHooks()
+    hooks: createHooks(),
   };
 
   // Register hooks
@@ -101,7 +135,7 @@ export async function build (rootDir: string, stub: boolean, inputConfig: BuildC
   await ctx.hooks.callHook("build:prepare", ctx);
 
   // Normalize entries
-  options.entries = options.entries.map(entry =>
+  options.entries = options.entries.map((entry) =>
     typeof entry === "string" ? { input: entry } : entry
   );
   const DEFAULT_EXTENSIONS = [".ts", ".tsx", ".mjs", ".cjs", ".js", ".jsx", ".json"];
@@ -163,19 +197,21 @@ export async function build (rootDir: string, stub: boolean, inputConfig: BuildC
   await ctx.hooks.callHook("build:before", ctx);
 
   // Start info
-  consola.info(chalk.cyan(`${options.stub ? "Stubbing" : "Building"} ${pkg.name}`));
+  consola.info(
+    chalk.cyan(`${options.stub ? "Stubbing" : "Building"} ${options.name}`)
+  );
   if (process.env.DEBUG) {
     consola.info(`${chalk.bold("Root dir:")} ${options.rootDir}
   ${chalk.bold("Entries:")}
-  ${options.entries.map(entry => "  " + dumpObject(entry)).join("\n  ")}
+  ${options.entries.map((entry) => "  " + dumpObject(entry)).join("\n  ")}
 `);
   }
 
   // Clean dist dirs
   if (options.clean) {
-    for (const dir of new Set(options.entries.map(e => e.outDir).sort())) {
+    for (const dir of new Set(options.entries.map((e) => e.outDir).sort())) {
       await rmdir(dir!);
-      await mkdirp(dir!);
+      await fsp.mkdir(dir!, { recursive: true });
     }
   }
 
@@ -206,11 +242,11 @@ export async function build (rootDir: string, stub: boolean, inputConfig: BuildC
   // Find all dist files and add missing entries as chunks
   const outFiles = await globby("**", { cwd: options.outDir });
   for (const file of outFiles) {
-    let entry = ctx.buildEntries.find(e => e.path === file);
+    let entry = ctx.buildEntries.find((e) => e.path === file);
     if (!entry) {
       entry = {
         path: file,
-        chunk: true
+        chunk: true,
       };
       ctx.buildEntries.push(entry);
     }
@@ -220,26 +256,62 @@ export async function build (rootDir: string, stub: boolean, inputConfig: BuildC
     }
   }
 
-  const rPath = (p: string) => relative(process.cwd(), resolve(options.outDir, p));
-  for (const entry of ctx.buildEntries.filter(e => !e.chunk)) {
+  const rPath = (p: string) =>
+    relative(process.cwd(), resolve(options.outDir, p));
+  for (const entry of ctx.buildEntries.filter((e) => !e.chunk)) {
     let totalBytes = entry.bytes || 0;
     for (const chunk of entry.chunks || []) {
-      totalBytes += ctx.buildEntries.find(e => e.path === chunk)?.bytes || 0;
+      totalBytes += ctx.buildEntries.find((e) => e.path === chunk)?.bytes || 0;
     }
-    let line = `  ${chalk.bold(rPath(entry.path))} (` + [
-      entry.bytes && `size: ${chalk.cyan(prettyBytes(entry.bytes))}`,
-      totalBytes !== entry.bytes && `total size: ${chalk.cyan(prettyBytes(totalBytes))}`,
-      entry.exports?.length && `exports: ${chalk.gray(entry.exports.join(", "))}`
-    ].filter(Boolean).join(", ") + ")";
+    let line =
+      `  ${chalk.bold(rPath(entry.path))} (` +
+      [
+        totalBytes && `total size: ${chalk.cyan(prettyBytes(totalBytes))}`,
+        entry.bytes && `chunk size: ${chalk.cyan(prettyBytes(entry.bytes))}`,
+        entry.exports?.length &&
+          `exports: ${chalk.gray(entry.exports.join(", "))}`,
+      ]
+        .filter(Boolean)
+        .join(", ") +
+      ")";
     if (entry.chunks?.length) {
-      line += "\n" + entry.chunks.map((p) => {
-        const chunk = ctx.buildEntries.find(e => e.path === p) || {} as any;
-        return chalk.gray("  └─ " + rPath(p) + (chunk.bytes ? ` (${prettyBytes(chunk?.bytes)})` : ""));
-      }).join("\n");
+      line +=
+        "\n" +
+        entry.chunks
+          .map((p) => {
+            const chunk =
+              ctx.buildEntries.find((e) => e.path === p) || ({} as any);
+            return chalk.gray(
+              "  └─ " +
+                rPath(p) +
+                chalk.bold(chunk.bytes ? ` (${prettyBytes(chunk?.bytes)})` : "")
+            );
+          })
+          .join("\n");
+    }
+    if (entry.modules?.length) {
+      line +=
+        "\n" +
+        entry.modules
+          .filter((m) => m.id.includes("node_modules"))
+          .sort((a, b) => (b.bytes || 0) - (a.bytes || 0))
+          .map((m) => {
+            return chalk.gray(
+              "  📦 " +
+                rPath(m.id) +
+                chalk.bold(m.bytes ? ` (${prettyBytes(m.bytes)})` : "")
+            );
+          })
+          .join("\n");
     }
     consola.log(entry.chunk ? chalk.gray(line) : line);
   }
-  console.log("Σ Total dist size (byte size):", chalk.cyan(prettyBytes(ctx.buildEntries.reduce((a, e) => a + (e.bytes || 0), 0))));
+  console.log(
+    "Σ Total dist size (byte size):",
+    chalk.cyan(
+      prettyBytes(ctx.buildEntries.reduce((a, e) => a + (e.bytes || 0), 0))
+    )
+  );
 
   // Validate
   validateDependencies(ctx);
@@ -251,9 +323,14 @@ export async function build (rootDir: string, stub: boolean, inputConfig: BuildC
   consola.log("");
 
   if (ctx.warnings.size > 0) {
-    consola.warn("Build is done with some warnings:\n\n" + [...ctx.warnings].map(msg => "- " + msg).join("\n"));
+    consola.warn(
+      "Build is done with some warnings:\n\n" +
+        [...ctx.warnings].map((msg) => "- " + msg).join("\n")
+    );
     if (ctx.options.failOnWarn) {
-      consola.error("Exiting with code (1). You can change this behavior by setting `failOnWarn: false` .");
+      consola.error(
+        "Exiting with code (1). You can change this behavior by setting `failOnWarn: false` ."
+      );
       // eslint-disable-next-line unicorn/no-process-exit
       process.exit(1);
     }