+++ src/utils.ts
@@ -1,26 +1,25 @@
 import fsp from "node:fs/promises";
-import { promisify } from "node:util";
 import { readdirSync, statSync } from "node:fs";
 import { dirname, resolve } from "pathe";
-import mkdirp from "mkdirp";
-import _rimraf from "rimraf";
 import jiti from "jiti";
-import consola from "consola";
+import { consola } from "consola";
 import type { PackageJson } from "pkg-types";
 import { autoPreset } from "./auto";
 import type { BuildPreset, BuildConfig, BuildContext } from "./types";
 
-export async function ensuredir (path: string) {
-  await mkdirp(dirname(path));
+export async function ensuredir(path: string) {
+  await fsp.mkdir(dirname(path), { recursive: true });
 }
 
-export function warn (ctx: BuildContext, message: string) {
-  if (ctx.warnings.has(message)) { return; }
+export function warn(ctx: BuildContext, message: string) {
+  if (ctx.warnings.has(message)) {
+    return;
+  }
   consola.debug("[unbuild] [warn]", message);
   ctx.warnings.add(message);
 }
 
-export async function symlink (from: string, to: string, force: boolean = true) {
+export async function symlink(from: string, to: string, force = true) {
   await ensuredir(to);
   if (force) {
     await fsp.unlink(to).catch(() => {});
@@ -28,23 +27,27 @@ export async function symlink (from: string, to: string, force: boolean = true) 
   await fsp.symlink(from, to, "junction");
 }
 
-export function dumpObject (obj: Record<string, any>) {
-  return "{ " + Object.keys(obj).map(key => `${key}: ${JSON.stringify(obj[key])}`).join(", ") + " }";
+export function dumpObject(obj: Record<string, any>) {
+  return (
+    "{ " +
+    Object.keys(obj)
+      .map((key) => `${key}: ${JSON.stringify(obj[key])}`)
+      .join(", ") +
+    " }"
+  );
 }
 
-export function getpkg (id: string = "") {
+export function getpkg(id = "") {
   const s = id.split("/");
   return s[0][0] === "@" ? `${s[0]}/${s[1]}` : s[0];
 }
 
-const rimraf = promisify(_rimraf);
-
-export async function rmdir (dir: string) {
+export async function rmdir(dir: string) {
   await fsp.unlink(dir).catch(() => {});
-  await rimraf(dir);
+  await fsp.rm(dir, { recursive: true, force: true }).catch(() => {});
 }
 
-export function listRecursively (path: string) {
+export function listRecursively(path: string) {
   const filenames = new Set<string>();
   const walk = (path: string) => {
     const files = readdirSync(path);
@@ -62,7 +65,7 @@ export function listRecursively (path: string) {
   return [...filenames];
 }
 
-export function tryRequire (id: string, rootDir: string = process.cwd()) {
+export function tryRequire(id: string, rootDir: string = process.cwd()) {
   const _require = jiti(rootDir, { interopDefault: true, esmResolve: true });
   try {
     return _require(id);
@@ -74,7 +77,7 @@ export function tryRequire (id: string, rootDir: string = process.cwd()) {
   }
 }
 
-export function tryResolve (id: string, rootDir: string = process.cwd()) {
+export function tryResolve(id: string, rootDir: string = process.cwd()) {
   const _require = jiti(rootDir, { interopDefault: true, esmResolve: true });
   try {
     return _require.resolve(id);
@@ -86,7 +89,10 @@ export function tryResolve (id: string, rootDir: string = process.cwd()) {
   }
 }
 
-export function resolvePreset (preset: string | BuildPreset, rootDir: string): BuildConfig {
+export function resolvePreset(
+  preset: string | BuildPreset,
+  rootDir: string
+): BuildConfig {
   if (preset === "auto") {
     preset = autoPreset;
   } else if (typeof preset === "string") {
@@ -98,7 +104,11 @@ export function resolvePreset (preset: string | BuildPreset, rootDir: string): B
   return preset as BuildConfig;
 }
 
-export function inferExportType (condition: string, previousConditions: string[] = [], filename = ""): "esm" | "cjs" {
+export function inferExportType(
+  condition: string,
+  previousConditions: string[] = [],
+  filename = ""
+): "esm" | "cjs" {
   if (filename) {
     if (filename.endsWith(".d.ts")) {
       return "esm";
@@ -111,10 +121,12 @@ export function inferExportType (condition: string, previousConditions: string[]
     }
   }
   switch (condition) {
-    case "import":
+    case "import": {
       return "esm";
-    case "require":
+    }
+    case "require": {
       return "cjs";
+    }
     default: {
       if (previousConditions.length === 0) {
         // TODO: Check against type:module for default
@@ -126,19 +138,37 @@ export function inferExportType (condition: string, previousConditions: string[]
   }
 }
 
-export type OutputDescriptor = { file: string, type?: "esm" | "cjs" }
+export type OutputDescriptor = { file: string; type?: "esm" | "cjs" };
 
-export function extractExportFilenames (exports: PackageJson["exports"], conditions: string[] = []): OutputDescriptor[] {
-  if (!exports) { return []; }
+export function extractExportFilenames(
+  exports: PackageJson["exports"],
+  conditions: string[] = []
+): OutputDescriptor[] {
+  if (!exports) {
+    return [];
+  }
   if (typeof exports === "string") {
     return [{ file: exports, type: "esm" }];
   }
-  return Object.entries(exports)
-    // Filter out .json subpaths such as package.json
-    .filter(([subpath]) => !subpath.endsWith(".json"))
-    .flatMap(
-      ([condition, exports]) => typeof exports === "string"
-        ? { file: exports, type: inferExportType(condition, conditions, exports) }
-        : extractExportFilenames(exports, [...conditions, condition])
-    );
+  return (
+    Object.entries(exports)
+      // Filter out .json subpaths such as package.json
+      .filter(([subpath]) => !subpath.endsWith(".json"))
+      .flatMap(([condition, exports]) =>
+        typeof exports === "string"
+          ? {
+              file: exports,
+              type: inferExportType(condition, conditions, exports),
+            }
+          : extractExportFilenames(exports, [...conditions, condition])
+      )
+  );
+}
+
+export function arrayIncludes(arr: (string | RegExp)[], searchElement: string) {
+  return arr.some((entry) =>
+    entry instanceof RegExp
+      ? entry.test(searchElement)
+      : entry === searchElement
+  );
 }